Extract AlarmCreationScreen and use for both Create and Edit. Refactor code.
