+++ app/src/main/java/com/example/alarmscratch/ui/alarmlist/composable/AlarmApp.kt
@@ -7,7 +7,10 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.alarmscratch.ui.alarmcreation.AlarmCreationScreen
-import com.example.alarmscratch.ui.navigation.Destination
+import com.example.alarmscratch.ui.alarmedit.AlarmEditScreen
+import com.example.alarmscratch.ui.navigation.AlarmCreation
+import com.example.alarmscratch.ui.navigation.AlarmDefaultSettings
+import com.example.alarmscratch.ui.navigation.AlarmEdit
 import com.example.alarmscratch.ui.navigation.NavigableScreen
 import com.example.alarmscratch.ui.settings.AlarmDefaultsScreen
 
@@ -16,18 +19,27 @@ fun AlarmApp() {
     val navHostController = rememberNavController()
     NavHost(
         navController = navHostController,
-        startDestination = Destination.NavigableScreen.route
+        startDestination = NavigableScreen.route
     ) {
-        composable(route = Destination.NavigableScreen.route) {
+        composable(route = NavigableScreen.route) {
             NavigableScreen(rootNavHostController = navHostController)
         }
-        composable(route = Destination.AlarmCreation.route) {
+        composable(route = AlarmCreation.route) {
             AlarmCreationScreen(
                 navHostController = navHostController,
                 modifier = Modifier.fillMaxSize()
             )
         }
-        composable(route = Destination.AlarmDefaultSettings.route) {
+        composable(
+            route = AlarmEdit.routeWithArgs,
+            arguments = AlarmEdit.args
+        ) {
+            AlarmEditScreen(
+                navHostController = navHostController,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+        composable(route = AlarmDefaultSettings.route) {
             AlarmDefaultsScreen(modifier = Modifier.fillMaxSize())
         }
     }
